+++ src/modules/method/index.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var glob      = require('glob');
 var logger    = require('yocto-logger');
 var _         = require('lodash');